boundary_layer_code/requirements.txt
imageio==2.35.1
matplotlib==3.4.3
numpy==1.19.2
tensorflow==2.4.0
Pillow
gdown==5.2.1
